requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.49.0
python-dotenv==1.0.1
lxml==5.3.0
